Include conflicting files in status output
This fixes #112 and fixes #110, by updating status to support any
conflicting paths. This allows someone to programmatically retrieve
those paths, rather than having to fall back to JGit.

This change exposed a "bug" where MergeOp was documented to abort
the merge on fail, leaving the repo in a clean state. This was not
accurate and the tests on succeeded because the status was incomplete.
To be more similar to normal Git behavior, this "bug" was kept and the
tests and documentation are updated to reflect the behavior (that was
already there).