Fix username-only creds
Username-only creds are useful for things like GitHub access tokens.
Prior to this commit, specifying only a username resulted in a null
password, which caused an ambiguous method call to the
UserNamePasswordCredentialsProvider constructor.

Now Credentials has been beefed up to fallback to empty strings for
values that aren't provided. There's also an explicit method for how to
check the validity of a Credentials object rather than needing to
specify it directly in each case.

Fixes #153