Treat credentials as immutable within repository
In 1.9.1, Credentials was changed from Immutable to Canonical in order
to better support default values that won't cause problems later on in
the code. Repository relied on Credentials being considered Immutable
and thus caused an exception anytime you pass credentials in to an open
(or similar call) that creates a Repository.

Instructing Repository to treat it as immutable should resolve this
issue #163.