Safer Jsch agent selection
When trying to select an agent to use for Jsch, you can get some general
exceptions or actual errors (NoClassDef or UnsatisfiedLinkError) to make
sure these more safely fail and allow Grgit to proceed to the next
option, catch Throwable instead of trying to identify specific errors.

This should resolve #171.