Remove custom exception type
Since JGit's exceptions are more information, all uses of custom
exception types were removed and JGit's are allowed to bubble up, when
there isn't any explicit handling needed.

This is not quite what the original requester in #146 had in mind, but this
seems like a better experience when something fails. For someone trying
to catch excpetions, they'll probably need to catch raw Exception. It
was unlikely trying to wrap them all was going to be successful anyway.