Throw exception if any ref fails to push
Pushes don't throw exceptions in JGit when a pre-receive hook declines a
push or any other reason the push was rejected logically. This makes it
difficult for Grgit users who run a push and then don't realize that
their updates didn't happen.

This new approach checks all of the push results and if any ref failed
to update, an exception is thrown.

This fixes #216.