Use static JGit version
Unfortunately nebula-maven-resolved-dependencies doesn't seem to put
static versions into the POM of com.gradle-plugin-publish artifacts. It
does affect everything I'm pushing to Bintray though. So we've been
using super permissive dependency ranges out to everyone and now, people
are being bit by a breaking change in JGit 5.

This is a quick fix for #229 to just make the JGit dependency static.