Replace Consumer with HasImplicitReceiver support
While Gradle usage from Groovy build scripts can use the Closure methods
and get a clean delgation:

  Grgit.open { currentDir = 'somewhere' }

From Kotlin build scripts, they fell back to the Consumer variant, but
that required the receiver (roughly equivalent in this context to the
Groovy delegate) to be explicitly referenced:

  Grgit.open { it.currentDir = 'somewhere' }

Using an interface with @HasImplicitReceiver allows for use like the
above Groovy usage.

Gradle provides the Action interface to meet this need, but in order to
preserve some independence from Gradle's API, since grgit is meant to be
generically a Groovy Git binding, we can't have the Action class or the
HasImplicitReceiver annotation on the classpath during compile.

We're doing a workaround at the moment that adds the annotation as part
of an AST transformation.

This change was provided by Alex Saveau.

I don't intend Kotlin usage to be a core supported feature of this
project, so this is just a best effort for now. If it ever requires more
extensive changes or breaks, I will favor removing it to making any
significant changes.