Expose grgit-core as compile dep of grgit-gradle
As of Gradle 5.0, Maven publications do not expose implementation
dependencies on the compile classpath of their published POM files.
Since consumers of grgit-gradle almost definitely need to interact with
Grgit types, the grgit-core dependency should be in the api
configuration.

Needed to utilize the java-library plugin to support this.

This fixes #267.