Close object reader used to parse commits
In converting all of the code to Java in another branch, this is the one
place I found an unclosed resource. This is a likely cause of the issues
with files being left open in Gradle builds. Will have to test this out
to confirm it resolves that issue, but it's at least a good change on
its own.

This is related to #182.